requirements.txt
# OMNIBOT v2.7 Titan - Requirements

# Web Framework
flask>=2.0.0
flask-socketio>=5.0.0
flask-session>=0.4.0

# Broker APIs
alpaca-py>=0.8.0
python-binance>=1.0.16
ib-insync>=0.9.70

# Data Processing
numpy>=1.21.0
pandas>=1.3.0

# Utilities
requests>=2.26.0
python-dateutil>=2.8.0
